rename State report controller to simply State
Now when we don't have configuration controller for the State,
we can make the name simpler.
