--- pkg/controller/add_nodenetworkstatereport.go
@@ -1,10 +1,0 @@
-package controller
-
-import (
-	"github.com/nmstate/kubernetes-nmstate/pkg/controller/nodenetworkstatereport"
-)
-
-func init() {
-	// AddToManagerFuncs is a list of functions to create controllers and add them to a manager.
-	AddToManagerFuncs = append(AddToManagerFuncs, nodenetworkstatereport.Add)
-}
